+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added support for plotting DRT results as gamma vs f.
   - Added CLI argument for plotting DRT results as gamma vs f.
 - Added circuits for generating mock data (`CIRCUIT_13`, `CIRCUIT_13_INVALID`, `CIRCUIT_14`, `CIRCUIT_14_INVALID`, `CIRCUIT_15`, and `CIRCUIT_15_INVALID`).
+- Added support for the modified Akima spline (`makima`) to the interpolation phase of Z-HIT and set it as the default.
 - Updated the TR-RBF implementation to be based off of a newer version of pyDRTtools.
 - Updated plotting functions to support drawing smoother lines if the input result has a `circuit` property.
 - Updated an exception message to provide more information if a variable that is being fitted is out of bounds.